(#115) Move some classes and packages
