+++ src/main/java/org/fusionsoft/database/snapshot/objects/dbms/postgres/PgFunctions.java
@@ -21,7 +21,7 @@ import org.fusionsoft.database.snapshot.DbObject;
 import org.fusionsoft.database.snapshot.Objects;
 import org.fusionsoft.database.snapshot.dbms.DbmsVersionOfConnection;
 import org.fusionsoft.database.snapshot.objects.ObjectsOfScalar;
-import org.fusionsoft.database.snapshot.objects.dbms.FunctionOfResultSet;
+import org.fusionsoft.database.snapshot.objects.resultset.FunctionOfResultSet;
 import org.fusionsoft.database.snapshot.query.Query;
 import org.fusionsoft.database.snapshot.query.pg.PgFunctionsQuery;
 import org.fusionsoft.lib.collection.ListOfResultSet;
