+++ src/main/java/org/fusionsoft/database/snapshot/objects/dbms/postgres/PgViews.java
@@ -19,7 +19,7 @@ import java.sql.Connection;
 import org.fusionsoft.database.snapshot.DbObject;
 import org.fusionsoft.database.snapshot.Objects;
 import org.fusionsoft.database.snapshot.objects.ObjectsOfScalar;
-import org.fusionsoft.database.snapshot.objects.dbms.resultset.ViewOfResultSet;
+import org.fusionsoft.database.snapshot.objects.resultset.ViewOfResultSet;
 import org.fusionsoft.database.snapshot.query.pg.PgViewsQuery;
 import org.fusionsoft.lib.collection.ListOfResultSet;
